rename: signIn 관련 클래스 auth -> users로 이동
