+++ src/main/java/synapse/dementia/domain/users/domain/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package synapse.dementia.domain.auth.domain;
+package synapse.dementia.domain.users.domain;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -7,9 +7,6 @@ import java.util.List;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-
-import synapse.dementia.domain.users.domain.Role;
-import synapse.dementia.domain.users.domain.Users;
 
 public class CustomUserDetails implements UserDetails {
